Implement Composition of graph computations
Summary:
Implements composition of two computations.
The composition is currently implemented such that given a base composed computation,
and a composer that returns a set of new keys given the key and result of the base computation,
returns a computation that takes in a composed key of the first computation and the final result class.

This lets you easily chain together several computations and request specific results given the very base key

Reviewed By: sbalabanov

fbshipit-source-id: 736824af6b