Generate buckconfig aliases automatically
Summary:
Right now we have a bunch of aliases for links to various buckconfig options. We do this because closure templates are cumbersome when trying to use them inline (you have to do `{call foo.bar}{param section: 'section'/}{param entry: 'entry'/}{/call}`, which is much longer than `{call buckconfig.foo_bar}`). However, these aliases have to be created by hand, and are slightly inconsistent.

Instead, just generate them from buckconfig.soy so that one can always guess the name without having to go verify that the sections are actually present. This runs whenever you start up soyweb-*.sh, and it has a test to ensure that it's not out of sync.

Reviewed By: styurin

fbshipit-source-id: f328c4bc46