Add script to alphabetize buckconfig within sections
Summary: We'd like to get some tech debt cleaned up around alphabetizing our buckconfig docs. Add a script to allow this to be done one section at a time so that changes are more atomic and reviewable. Also allow certain entries to be excluded and kept where they are. This will likely be removed down the road once the work is done, though we could also use the script to ensure that things are properly alphabetized until we have a plan on how we want to auto generate this documentation.

Reviewed By: styurin

fbshipit-source-id: 1a954ab21c