Improve ExcludeFromRuleKey
Summary:
When something is excluded from a rulekey, the normal
serialization/input handling won't occur (serialization will throw an
exception and input handling will ignore it).

Just doing this silently isn't great, we should force people to
explicitly specify the behavior.

This adds a couple arguments to ExcludeFromRuleKey that are used to
specify the serialization and input handling behavior. These are
required.

This requires some decent sized updates to the handling of
CustomFieldBehavior to be able to get it out of the ExcludeFromRuleKey
annotation, but it's mostly not too bad.

Updates the few uses of ExcludeFromRuleKey to specify these new things
(and delete the corresponding CustomFieldBehavior if present).

fbshipit-source-id: 94dc4379ab