Add smarter Java lookup logic to Python wrapper
Summary:
Try to find the right Java path based on the Java major version we were compiled against. The new Java lookup logic is:
1. If `JAVA_HOME` is set, use that, unless it looks like a path for the wrong version of Java based on regex match
2. Otherwise, try to use a known Java path for the required Java version
3. Fall back to PATH if all else fails

This behavior applies to Java 8 buck as well, so users should occasionally benefit from this right away, but the main point is to ease the Java 11 migration when callers of Buck have `JAVA_HOME` set to Java 8, or only have Java 8 on the path.

Reviewed By: sbalabanov

fbshipit-source-id: 9e4e9ae2db