Update TargetEvaluator in MultitenantQueryEnvironment to support "//..."
Summary:
Fortunately, adding special support to map `//...` to
`Index.getTargets()` falls out fairly naturally from the
`QueryEnvironment` design. This is great because (although we do not
do this yet!) it is straightforward for us to implement
`Index.getTargets()` using multiple threads to speed it up.

I will support wildcard patterns more generally in a subsequent diff.

Reviewed By: sbalabanov

fbshipit-source-id: a7735f74ea