Introduce FsAgnosticPath for multitenant packages.
Summary:
For multitenant, we like `java.nio.Path` because it is safer than using
`String`. However, it also has downsides:

* It has an association with `java.nio.file.FileSystem` that is not appropriate
* Its `toString()` uses `\\` on Windows, which has necessitated the introduction of `Index.pathWithUnixSeparators()`

This introduces `FsAgnosticPath` that should be used in place of
`java.nio.Path` within `com.facebook.buck.multitenant`, which should
address these issues.

`FsAgnosticPath` is a simple wrapper around a `String`; however, it must
go through `FsAgnosticPath.of()` for construction, so we ensure that it
represents a normalized, relative path. (If we want to get more
precise in the future, we might want to change the underlying
representation to `byte[]`, as a path has no default encoding.)

We also create a `Cache<String, FsAgnosticPath>` with soft values in
an attempt to save memory. We will have to see whether this is actually
advantageous, in practice.

Reviewed By: sbalabanov

fbshipit-source-id: 0a76240590