Improve `buck audit mbr isolation`
Summary:
This `buck audit mbr isolation` to group errors by package and then
print them that way.

Errors are now grouped by:

1. rule type
2. error message
3. package

And printed in order of most instances.

The biggest benefit is that when an issue has to be fixed in the repo,
rather than in buck itself, it becomes possible to identify the most
important packages to convert.

Reviewed By: rajyengi

fbshipit-source-id: 09025ca9a8