StringWithMacrosConverter: avoid generating one-item `CompositeArg`s
Summary:
The `StringWithMacrosConverter` takes a `StringWithMacros` and converts it into
a `Arg`.  As the latter can include multiple chunks of strings and macros, it
converts each, individually, to an `Arg` and then glues them together with a
`CompositeArg`.  However, in the vast majority of cases, `StringWithMacros`
contain only a single string element (as `*_flags` parameters typically don't
use macros).  In this case, we're unnecessarily allocating a `ImmutableList` and
and `CompositeArg` instance to wrap it.  This has more significant impliciations
for rule key caches, as the `CompositeArg` is independently cached and so
causes bookkeeping overhead.

This diff instead detects the case when the `StringWithMacros` contains only a
single macro or string, and avoids the `CompositeArg` in this case.

Reviewed By: cjhopman

fbshipit-source-id: 223cc2d20b