StringWithMacrosArgConverter: add `ActionGraphBuilder`
Summary:
This was originally added via a parameter in D8081398, as some places capture
and stored a `StringWithMacrosArgConverter`.  I couldn't find any cases of these
in a quick audit, so this may be safe now.

If so, this allows most downstream users to avoid an additional lambda
allocation which captures the `ActionGraphBuilder` object anyway (in most cases,
the `StringWithMacrosArgConverter::convert` is used in a stream pipeline).

Reviewed By: jtorkkola

fbshipit-source-id: 8fe7007873