Make generation the primary argument to Index methods instead of commit.
Summary:
Effectively, all of the public methods of `Index` that took a commit
would start by taking the `rwLock` to call `commitToGeneration[commit]`.
For repeated calls into `Index` with the same `Commit`, this seems
unnecessarily wasteful (as `MultitenantQueryEnvironment` needs to do),
so this changes the API of `Index` to take a `Generation` rather than a
`Commit` as the primary parameter for its public methods.

As such, `MultitenantQueryEnvironment` has been reworked to take a
`Generation` rather than a `Commit` for use with its private `Index`
field.

Reviewed By: bobyangyf

fbshipit-source-id: dfb2aa1f88