Support static-pic,prof for Haskell
Summary:
We added support for static-pic,prof for Haskell in this change.
Now we have

* `-fexternal-interpreter` to support TemplateHaskell
* `-fexternal-dynamic-refs` to support `static-pic`

we can also clean up some hacky implementation in buck:

* `static-pic` should not link to `.so` (should not use `-dynamic`)
* both `static,prof` and `static-pic,prof` are supported now
* ghci will always use `static-pic` or `static-pic,prof`
* `shared,prof` is still not supported

We should not include vanilla libraries in profiling package, but some
use cases currently depend on this and we need migrate them first.

Note: this change requires ghc-8.6 or newer version.

Reviewed By: josefs, styurin

fbshipit-source-id: 67ba4af5b2