Fix BuildLogHelper to handle proper file attributes when walking directories
Summary:
`Files.isSymbolicLink` is hiding `IOException` which means if a file doesn't exist it returns `false`. At the same time, the call in `BuildLogHelper` was used with a relative path which was resolved against the location where buck is running, not the root of project filesystem. The fix is to use attributes passed to the callback.

Another change is to actually handle exceptions when they happen.

Reviewed By: philipjameson

fbshipit-source-id: fa942e66a8