Remove IndexReadLock: do synchronization internal to methods.
Summary:
This pattern did not work out the way I expected: it seems cleaner (and
likely more efficient) for each public method of `Index` to hold the
read lock for the brief period it needs it rather than for the entire
method.

This will also pave the way to create a separate object (with its own
lock) that contains both `buildPackageMap` and `ruleMap`. I suspect
this will make it easier to implement support for local changes from
the user because then we can have a "decorator" `Index` with its own
version of this new, separate object that delegates to the original,
as appropriate.

Reviewed By: bobyangyf

fbshipit-source-id: f77d4440c8