Have WatchmanWatcher read directory changes from Watchman
Summary:
This change is a preparation to turn Watchman to also report changes in directories to Buck. At this moment Buck only gets notified when a file has changed.

This change also aggregates all Watchman results and send them in one event, to better utilize invalidation strategies down the stack.

Reviewed By: bobyangyf

fbshipit-source-id: 7a61176b67