Introduce BuildTargets.kt for multitenant service.
Summary:
This consolidates a bunch of one-off build target parsing logic that
was strewn about the multitenant code and puts it in one place.
Further, it now uses the existing logic in `BuildTargetPatternDataParser`
rather than rolling its own.

This made it possible to eliminate `InternalTypeMapper`, which simplifies
`Index` and `IndexAppender`. I think this sets things up well so that
incorporating local changes from a user can be handled by `IndexAppender`,
in that it will use `determineDeltas()` to produce new deltas, which can
then be used to create a new `IndexGenerationData` that composes the
existing one in `IndexAppender`. Using that, we can create a new `Index`
with this one-off `IndexGenerationData`, pass it to a new
`MultitenantQueryEnvironment`, and be on our way!

Reviewed By: sbalabanov

fbshipit-source-id: 545e2df7f7