Introduce ForwardingGenerationMap.
Summary:
The Javadoc for `ForwardingGenerationMap` explains this in more detail,
but basically, this provides a `GenerationMap` that provides a view of
an existing `GenerationMap` with a set of deltas superimposed for a
specific generation. This will be a key abstraction in implementing
support for local changes in the multitenant service.

Reviewed By: sbalabanov

fbshipit-source-id: 69c271dffd