Derive Index that incorporates local user changes.
Summary:
This introduces `Index.createIndexForGenerationWithLocalChanges()`,
which returns a (possibly new) `Index` that presents a view of
the index data at the specified generation with the specified local
changes taken into account. This is the key functionality that makes
this a truly multitenant service!

Note this moves the `determineDeltas()` from `IndexAppender` into its
own `Deltas.kt` file because now both `Index` and `IndexAppender`
need to use this logic. There is also some refactoring in
`IndexGenerationData` so that we can derive a new `IndexGenerationData`
from an existing one using `ForwardingGenerationMap`s to blend the
local changes with those from trunk while using the same locks to guard
the collections in both the existing and new `IndexGenerationData`
objects.

Reviewed By: sbalabanov

fbshipit-source-id: a081ef85f4