Remove unused dependency in multitenant on a mainline Buck testutil.
Summary:
Currently, I am trying to use a different test utility from the Buck
module in Intellij (`FakeProjectFilesystem`), but IntelliJ does not
seem to recognize it. I think this is because `multitenant.iml` is not
configured properly, so I checked to see if the multitenant test code
was using any of Buck's test code. This was the only instance, but it
turned out it was unused, which is probably why I have not seen an issue.

Reviewed By: bobyangyf

fbshipit-source-id: 19de3830c2