Make Intellij not import inner classes
Summary:
This is usually not what we want, we have many inner classes that have
the same name as other inner classes and we use the outer class name to
differentiate them (ex. Builder, Type).

The codeStyleConfig change comes along with it because our old value was
incorrect and intellij wants to fix it for us (intellij seems to expect
only one of USE_PER_PROJECT_SETTINGS or PREFERRED_PROJECT_CODE_STYLE).

fbshipit-source-id: 4adcc77797