UnconfiguredBuildTargetParser - unified build target parser which is environment-agnostic
Summary: This parser simply parses a string cell//path/to:target into UnconfiguredBuildTarget. It is going to replace BuildTargetParser and satellites which has a lot of dependencies and also parse considering environment state (i.e., absolute cell paths). The new parser does not depend on running environment and can be used in multitenant applications.

Reviewed By: bobyangyf

fbshipit-source-id: 2ad1409835