cxx_toolchain: headers_whitelist
Summary: While forbidding whitelisting of untracked header files (and moving this logic to wrappers of cxx tools) is a good step in the long run, it is unrealistic for all the available toolchains to implement this logic before migrating to cxx_toolchain rules. Basically, a sole migration to cxx_toolchain is a good motivation, but the requirement to implement such wrappers first is a too hard requirement for now.

Reviewed By: cjhopman, styurin

fbshipit-source-id: c98eec66bd