Introduce Index.getTargetsInOwningBuildPackage()
Summary:
This builds on `getTargetsInBasePath()`, except that it takes
responsibility for doing the recursive step of finding the nearest
ancestor for the specified `basePath` when it does not immediately
contain a build file.

Note that we add this as its own method of `Index` rather than adding
a helper method built on top of `Index` so that we only take the lock
on the build package map once, even in the recursive case.

Reviewed By: sbalabanov

fbshipit-source-id: 12cb595a92