Change the contract for Index.getTransitiveDeps().
Summary:
Modified the contract so that it explicitly includes the specified
target in the set that is returned. It turns out that this method was
only used in unit tests, anyway, so it was easy to update.

Changing the method in this way will make it easier to modify so it
can be used to implement
`MultitenantQueryEnvironment.getTransitiveClosure()`.

Reviewed By: sbalabanov

fbshipit-source-id: 444b16596b