Change Index.getTransitiveDeps() to take multiple targets.
Summary:
This is more efficient than calling `getTransitiveDeps()` with multiple,
individual targets and merging the resulting sets.

While in this code, I decided to use a `HashSet` for the output because
unlike `kotlin.collections.Set`, it is possible to specify the
`initialCapacity` when creating the set, which seems like a good thing
to do since `getTransitiveDeps()` could return a very large set.

This diff also updates `MultitenantQueryEnvironment` to leverage
`Index.getTransitiveDeps()`. We do not verify this as part of this diff:
once we implement `getReverseDeps()`, we can add `AllPathsFunction`
and `RdepsFunction` to the list of functions that
`MultitenantQueryEnvironment` supports and will verify it then.

Reviewed By: sbalabanov

fbshipit-source-id: d7f33fa8d2