Use a persistent collection to store build rule names.
Summary:
The `buildPackageMap` in `IndexGenerationData` is used to store the set of
rule names defined in a build package as it changes over time. Previously,
every time this set changed, we would store a full copy of the new set. For a build
package with a large number of rules but few changes to the set, those full
copies are potentially wasteful.

To attempt to mitigate this issue, this diff introduces the use of
[persistent data structures](https://en.wikipedia.org/wiki/Persistent_data_structure)
so that multiple versions of a set can share memory rather than be
full, independent copies. Although there are many open source,
"persistent collections" libraries available, we are currently going with Vavr.

As we might want to switch to a different library in the future, we introduce
the use of a Kotlin `typealias` along with an extension method to make
such a migration a little easier. Admittedly, this approach does not hide any
of the API of `io.vavr.collection.Set` when `BuildRuleNames` is used,
but using a wrapper object would entail an extra allocation per set.

We take advantage of the persistent collection in `determineDeltas()`
where previously we would just create a new `Set<String>` from scratch
for a new `BuildPackageDelta.Updated` object, but now we derive the
new `Set<String>` from the old one so they can share memory.

Reviewed By: bobyangyf

fbshipit-source-id: 0281e18748