Split RuleDelta.Updated into Added and Modified.
Summary:
Now that we have a legitimate case in `Deltas.kt` handles "Added" and
"Modified" differently, create separate values for them in the
`RuleDelta` enum. `MapDiff` was already doing the work to distinguish
these two cases.

Incidentally, the `RuleDeltaUpdate` interface will likely go away in the
near term because when we add an index for reverse dependencies,
we will want to treat added vs. modified separately in the places where
those cases are now unified.

Reviewed By: bobyangyf

fbshipit-source-id: addf31933d