Introduce RdepsSet.
Summary:
RdepsSet represents a set of reverse dependencies for a build rule.
We try to be clever by implementing this as an algebraic data type where
it is backed by either `IntArray` or `io.vavr.collection.Set<Int>`.

The idea is that when the set of rdeps is small or changes infrequently,
an `IntArray` should be a memory-efficient way to store the rdeps.
However, if rdeps is large or changes frequently, we prefer to use a
persistent collection so that more memory can be shared across versions
of the set.

Admittedly, this comes at the cost of additional complexity, but this is
based on the hypothesis that for build rules with a large number of
rdeps (like Guava in a large Java codebase), this will be worth it.

Reviewed By: sbalabanov

fbshipit-source-id: 2f82292a98