Maintain a reverse dependency map as part of the multitenant service.
Summary:
This diff introduces the logic to maintain an inverse dependency map,
which can be queried via `Index.getReverseDeps()`.

This adds a new collection to `IndexGenerationData`: `rdepsMap`.
Like `buildPackageMap` and `ruleMap`, it has its own lock.
Note values in this map are of the new `RdepsSet` type introduced in
the previous diff.

Most of the new logic to keep `rdepsMap` updated is in `determineDeltas()`.
Specifically, `Deltas` gets a new field, `rdepsDeltas: Map<BuildTargetId, RdepsSet?>`,
which `determineDeltas()` is responsible for populating. As was the case
in `RdepsSet.kt`, there is a bit of logic used to compute diffs between
sorted lists of dependencies efficiently.

Once `rdepsDeltas` is constructed, both `IndexAppender.addCommitData()`
and `Index.createIndexForGenerationWithLocalChanges()` required very little
code to take advantage of this new information.

This diff also adds `deriveRdepsDeltas` to `RdepsSet.kt` to help map between
some of the internal data structures in `determineDeltas()` to something that
can be used for `Deltas.rdepsDeltas`.

In the end, the code to implement `Index.getReverseDeps()` is fairly simple
because it is primarily a lookup in the `rdepsMap`, the result of which are
build target ids that are converted to actual build targets.

Reviewed By: sbalabanov

fbshipit-source-id: 57db39af6d