Add support for rdeps() in the multitenant service.
Summary:
Unfortunately, a "show references" query like the following will be very
inefficient in the multitenant service even though it doesn't have to be:

```
rdeps(//..., //java/com/example:A, 1)
```

The inefficiency comes from `//...` as the "universe" argument, which
will be a set with all of the build targets in the project. By the time
`RdepsFunction` is invoked, `//...` has already been evaluated, so there
is no way to avoid this since we don't have anything like a
"query optimizer" powering `buck query` today.

I think I am going to introduce a new function named `refs` such that
the following would be equivalent to the above:

```
refs(//java/com/example:A)
```

This would make it efficient for the multitenant service to power a
"show references" query from an IDE (likely as part of an LSP).

Reviewed By: sbalabanov

fbshipit-source-id: fd9f2936c6