Change json format for building multitenant index
Summary:
This makes it closer to what Buck produces by serializing RawTargetNode/RawTargetNodeWithDeps/RawTargetNodeWithDepsPackage

The main motivation for the change is that 'deps' in attributes and 'deps' in RawTargetNodeWithDeps are different and need to be handled differently. 'deps' in attributes is what user specified in build file, but 'deps' in RawTargetNodeWithDeps are all posible dependencies inferred by Buck from attributes.deps and also other sources too.

Reviewed By: bolinfest

fbshipit-source-id: b82f66d9e8