Convert BuildLogEntry to @BuckStyleValue
Summary:
- migrates to the new style of immutable
- replaces `builder()` usage with constructor calls
- replaces abstract class for an interface

file `BuildLogHelper.java` features a bigger refactoring than last commits. The issue was to transform *builder* code to a constructor, but the old code contained multiple conditional paths.

the refactoring strategy was to flatten the `ifs`, and use `Optional` along with its functions where possible instead. Then create the values that will be given to the constructor at the end. Most of the conditions were there to check that values were present, so it's valuable to write it this way (I think).

this commit is the last one of the migration to the new Immutable style for package `doctor.config`

Reviewed By: bobyangyf

fbshipit-source-id: e69ef4691d