Introduce IndexAppender.getLatestCommit().
Summary:
The motivation is to make it possible for someone to make a query to
the service against the "latest" commit that it has in its index.
Because we return the latest commit (rather than just the latest
generation), we can also tell the caller the commit at which their
query was made.

This can also be used to query the service to see how far it is behind
`master`.

Reviewed By: sbalabanov

fbshipit-source-id: 01909cce28