Reset parser state when configuration targets change
Summary:
This change extends `DaemonicParserState` to invalidate parser state when there is a possible change in configuration rules.

Configuration rules can affect the target graph structure, but Buck target graph model doesn't contain information about dependencies on configuration rules. Adding this information would complicate the model and since changes to configuration rules are not frequent, it's reasonable to reset the whole parsing state when there is a change in configuration rules.

The approach in this change is based on detecting changes in build files that contain configuration rules. In situations when a build file contains both build targets and configuration targets this can result in unnecessary invalidations when only build targets are changed. This repository configuration is unusual and it's recommended to keep configuration rules separately from build rules so in practice this should not be a problem.

Reviewed By: sbalabanov

fbshipit-source-id: 1f6e691328