Fix RE event ordering
Summary:
Currently, we are sending events in this order:
materializing_start
deleting_stale_start
deleting_stale_finish
materializing_finish

The listener for these events doesn't really expect the states to be
nested in this way, and the real work for materializing is happening in
an unexpected place (after deleting_stale_finish).

Reorder this to be better.

Reviewed By: rajyengi

fbshipit-source-id: dbd71bb675