Add FastPaths.hashPathFast
Summary:
FastPaths provides utilities to more efficient operate on Paths
considering that sometimes they are BuckUnixPath (which has vastly
different performance properties to normal Java Paths :().

FastPaths.hashPathFast iterates over the segments in the Path and adds
them to the Hasher. This allows us to hash a Path without calling (the
extremely slow and garbage-generating) BuckUnitPath.toString().

FastPaths access BuckUnixPath internals via
BuckUnixPath.InternalsForFastPaths to try to limit the extent that we
poke into its internals. It isn't just in the same package because
BuckUnixPath is loaded into the bootloader.

Reviewed By: sbalabanov

fbshipit-source-id: 725cb667f1