Support Path as a key for rulekey entries
Summary:
setReflectively is now only used from within the rulekey computation
framework itself, so we have a bit more freedom to unrestrict things.

Depfile (and manifest) computation for c++ adds a ton of Paths to
rulekeys as keys (and it needs to be as keys to get elided when not in
the depfile). Calling BuckUnixPath.toString() that many times is
ridiculously inefficient and generates a ton of garbage.

Allowing a Path as a key to setReflectively allows us to add those
without converting them to strings.

This shows ~1.6x speedup in depfile computation of a medium sized mixed
java/c++ project and ~1.5x speedup for manifest computation of the same.

Reviewed By: sbalabanov

fbshipit-source-id: 4dc69f4d54