Export Label() function
Summary: Export Label() from skylark into the skylark extension file parser. This reuses much of the built-in logic from Skylark with a few changes. We're re-using this here because there are a lot of assumptions built into skylark about the Label class being used. Many other built-ins from Bazel/Skylark will not be used directly as they're highly coupled to Bazel. Label is kind of the exception.

Reviewed By: bobyangyf

fbshipit-source-id: 6607014681