Optimize DebugPathSanitizer API
Summary:
Converting Path -> String is quite expensive. Previously, for every
sanitized flag, we were converting all the paths that need sanitization
to strings. Instead, only provide an API the returns a
Function<String,String> that can perform sanitization (and an api that
takes a list<string>) to help keep users on the fast path.

Reviewed By: sbalabanov

fbshipit-source-id: e962ad548e