Always prioritize strategy flag over experimental_strategy
Summary: If it is set, always use the "strategy" setting to choose the mode. Only if it is not set and user is in experiment/project is whitelisted, use "experimental_strategy". This allows users to have one clear way of overriding strategy/mode regardless of whether they are in experiment or not.

Reviewed By: rajyengi

shipit-source-id: 640a15fe33