Optimize LimitedFileHashCacheEngine
Summary:
Convert a `Preconditions.checkState(<expr>, "msg " + path)` to
`Preconditions.checkState(<expr>, "%s msg", path)`.

This is a rather big improvement, as the first one converts the path to
a string all the time (and the other never does in normal operation).
BuckUnixPath.toString is (1) rather expensive and (2) generates a bunch
of garbage.

This function is called a ton (basically x times for each class referenced in each
java file).

Reviewed By: sbalabanov

shipit-source-id: abaddc9d2f