Break up SkylarkProjectBuildFileParserTest
Summary: Break out some common functionality used in SkylarkProjectBuildFileParserTest so that it can be used in a new class that is dedicated to testing details for user defined rules. The general skylark parsing logic tests are already 1600 lines long, so it makes sense to make the tests for user defined rules, that may also become very long, their own test suite.

Reviewed By: bobyangyf

shipit-source-id: ac548df331