Add SkylarkUserRule
Summary: Add the actual callable that is returned from the rule() function in the parse environment. This is called by users to create instances of their target which are returned from the parser. While parameter names are checked here, argument types are not checked until type coercion later. This will eventually be used to create the SkylarkDescription/SkylarkDescriptionArgs

Reviewed By: bobyangyf

shipit-source-id: 25941cf72d