Export rule() function
Summary: Add and export the `rule()` function to the parser. At the moment it is not hooked into the backend, and calling it on a full build will cause errors, as the type-coercer is not hooked up. However, it will export the targets and their attributes from the parser properly so that future diffs can plumb the rules all of the way through.

Reviewed By: bobyangyf

shipit-source-id: cb44dea1da