Fix inefficiency in JarBuildStepsFactory
Summary:
This is using a list of paths as String objects and then converting them
to Path many (many) times. For normal Paths that's not great, but not
terrible since it's just an allocation, but for BuckUnixPath it requires
several allocations and interning of strings. This is a particularly hot
function (as it's used in depfile/manifest computation which doesn't do
quite the good caching that default/input-based keys do).

Reviewed By: jtorkkola

shipit-source-id: d480e4b3b2