Fix FileBundler logic
Summary:
FileBundler had some complicated logic that was trying to handle
multiple different logical cases with some shared function. That shared
function was making assumptions about the state of the world that just
aren't true and we were ending up with copy steps that just copy a file
onto itself (and "itself" was an output for a totally different rule
that this one shouldn't be writing to).

This changes the logic to much more explicitly handle the different
cases.

Reviewed By: sbalabanov

shipit-source-id: 2aab510141