Platform specific thinlto flags with tests
Summary:
[buck][incremental][thinlto]Platform specific thinlto flags with tests

Unfortunately we need to guarantee the linker in the toolchain to actually
run the tests in buck. For stock ld (not gold) on linux, or stock ld64 on
mac, we're going to crash on sending the relevant linker options.

I just bypassed the linker with a custom script and manually emitted the
correct filenames. The content of the files is not what the linker would
normally emit. The idea here is to check that the arguments have been
passed correctly, and the files written to the correct location.

Reviewed By: cjhopman

shipit-source-id: e0ad47cdad