Add support for Buck invocations with isolated daemon/cache/outputs
Summary:
This diff proposes some changes to allow Buck to run in an "isolated" mode - where outputs, intermediate files etc. are not written to the usual `buck-out` directory. Instead, they are written to a directory whose name is a user-supplied prefix plus `-buck-out` that sits next to where the traditional `buck-out` would go (i.e. at the repository root).

The change also means a separate `buckd` will be launched that is tied to this isolated output directory.

The motivation for the change is to support Nuclide Language Services (and potentially other clients too) that want to use Buck to harvest information about the source tree and build process, but do not want to conflict with user-initiated build/test commands.

A key principle I've tried to follow with the change is to not interfere with "normal" Buck usage in any way. So this diff should leave all existing Buck invocations and usage unaffected.

Reviewed By: bobyangyf

shipit-source-id: b68b8f211f