Fix memory leak in cxx toolchain.
Summary:
The buildRuleResolver is a weak key in a cache in DefaultLinkerProvider, and the values are instances of BuildRuleResolverCacheByTargetConfiguration. Because the latter has a strong reference to the key object, the buildRuleResolver objects were always retained. This resulted in a leak of around 650MB each time the build graph for fb4a was computed (and probably similarly sized leaks for other targets).

Fixed by making BuildRuleResolverCacheByTargetConfiguration.buildRuleResolver also a WeakReference. This should be safe, since Java guarantees that it atomically frees all weak references to an object.

Reviewed By: styurin

shipit-source-id: 9267cc6333