Rename NativeLinkable->NativeLinkableGroup
Summary:
We're going to more explicitly separate the parallel graphs for each c++
platform. To that end, I'll introduce a new NativeLinkable that
represents the information in the current one that's just for a single
platform.

That new NativeLinkable should be the default used one (and ideally we
eventually delete the current one that aggregates across multiple
platforms) and so I didn't want to give it a longer name.

Reviewed By: styurin

shipit-source-id: 1a1a240e6b