Fail go project when CL arguments resolves to empty set of targets
Summary:
This logic is not necessary since arguments are checked for being empty at the beginning of `parseTargetsAndRunProjectGenerator`. If arguments are empty it uses `//...` as an input which should return all the targets in the graph.

Checking for empty targets after processing arguments and parsing the whole repository means that Buck would parse the whole repository when input arguments are resolved into empty set of targets which counter-intuitive.

Reviewed By: jtorkkola

shipit-source-id: 366e2b9026