Fail Xcode project when CL arguments resolves to empty set of targets
Summary: This changes makes Xcode project generator consistent with other project generators in that sense that it uses `//...` as a target pattern when CL arguments are empty. This allows to remove some duplicated logic to handle the case with empty set of targets. This also means that Buck would no longer parse the whole repository when input arguments are resolved into empty set of targets.

Reviewed By: jtorkkola

shipit-source-id: d2e2273190