Add Build Report Upload
Summary:
A build report will basically replace most of the information that right now is only submitted with a `buck rage`. As a start, this diff adds only the configuration map to it. Next diffs are going to add source control information.
It is implemented as a background task so that it doesn't block the user's command, but ideally it would capture information before the background task runs, so that if the user changes something after the build it doesn't affect the report.

Reviewed By: philipjameson

shipit-source-id: bef5f1b489