Add write action
Summary:
Add a `write()` action. This follows the arguments in https://docs.bazel.build/versions/master/skylark/lib/actions.html#write

To test this more easily and with less boilerplate, also added a little helper that runs an action execution step with the given action. This was mostly to make sure that we're not being too synthetic in our creation of the action; there will be a properly constructed artifactfilesysten if we do it in the step, rather than creating one manually.

Reviewed By: bobyangyf

shipit-source-id: 6a52b818c4