Begin the separation of CxxLibrary by platform
Summary:
This introduces CxxLibraryDescriptionDelegate.ConfiguredDelegate. All
the functions from the previous delegate are moved to this new one and
replaced with a single method `requireDelegate(BuildTarget,
GraphBuilder, CxxPlatform)`. This allows the configured delegate to have
the platform and build target baked in.

This will be followed up with the changes need to split CxxLibrary in
the same way so that we actually get parallel graphs for different
platforms rather than ones that keep weaving back and forth between the
mega CxxLibrary rules carefully (or often not carefully) passing the
platform down.

Reviewed By: bobyangyf

shipit-source-id: 507d0b050f