Rename CxxLibrary -> CxxLibraryGroup
Summary:
The current CxxLibrary actually represents and exposes information for a
whole group of cxx libraries (for different platforms) that have
(possibly wildly) different behavior and are really logically part of
their own platform-specific graph of libraries.

We'd like to make that logical graph more explicit and introduce a new
thing to represent the platform-specific information about the library.
As the default should be to interact with the platform-specific
information, I'm going to use 'CxxLibrary' to represent the
platform-specific one and 'CxxLibraryGroup' to represent the
non-platform-specific one.

One might wonder, with the rename of
NativeLinkable->NativeLinkableGroup, why aren't other implementations
renamed? If you dig into the other implementations, they don't actually
represent multiple different platform-specific things.

Reviewed By: andrewjcg

shipit-source-id: 21f6b12da0