Strategy for auto enabling Remote Execution is now configurable
Summary:
The existing strategy is to always enable Remote Execution if either RE experiment is enabled or the project being built is whitelisted. This is not always desirable.

There is is now a new config option:

```
[remoteexecution]
auto_re_strategy=
```

Which can be used to vary this behavior. The settings are as follows:
- DISABLED: never auto enable RE
- RE_IF_EXPERIMENT_ENABLED: enable RE if experiment is enabled
- RE_IF_WHITELIST_MATCH: enable RE if current target matches whitelist (and user not blacklisted)
- RE_IF_EXPERIMENT_ENABLED_AND_WHITELIST_MATCH: enable RE if both of above conditions are true
- RE_IF_EXPERIMENT_ENABLED_OR_WHITELIST_MATCH: enable RE if one or more of RE_IF_EXPERIMENT_ENABLED or RE_IF_WHITELIST_MATCH is true (Note: this was the existing behavior and is now DEFAULT).

Reviewed By: michsien, rajyengi

shipit-source-id: 8af620a3a5