Correct test runner class loading for Java 11
Summary:
## Summary

The previous fix to test runner classloading for Java 9+ did not cover all cases. In particular, anything involving in-proc Java compilation in tests did not work.

## Problem

The problem is that in Java 9+, because we can't directly modify the system classloader anymore, we created a new classloader that was a child of the platform classloader.

However, in Java 9+, the compiler classes are loaded in the system classloader. In order for our classloader proxying to work, we need `BuckJavacTaskProxy` to be in the system classloader, i.e. in the real classpath for `FileClassPathRunner`, accessible by the classloader for `JavaCompiler.CompilationTask`. This is because `BuckJavacTaskProxy` requires `BuckJavacTaskProxyImpl`, created in the plugin loader (see `PluginLoader`/`PluginClassLoader`) , to be able to derive from `BuckJavacTaskProxy`, requiring accessibility from the plugin class loader's parent chain.

This is not an issue for Buck itself, as opposed to tests, as `ClassLoaderBootstrapper` puts everything into a classloader that is a child of the system classloader, satisfying the above requirement.

## New approach

The approach I decided on for tests is to leverage the new [argfile support](https://docs.oracle.com/en/java/javase/11/tools/java.html#GUID-4856361B-8BFD-4964-AE84-121F5F6CF111) for long classpath arguments when we think we're launching at least Java 9, and leaving the logic unchanged otherwise. This means we now have 3 modes in `FileClassPathRunner`.

For Java 9+, when Buck is aware that it's launching Java 9+, this approach ensures that everything goes into the system classloader, which is exactly what happens for Java 8-, satisfying the plugin loader's requirement.  However, as of yet, Buck doesn't always know when it's launching Java 9+, so we need the old approach to handle when Buck launches a target that targets Java 8-, but `java` on the path turns out to be 9+. See the comments in `FileClassPathRunner`.

The long term plan is to make Buck smarter about Java invocations, such that it looks up the right Java binary path based on the version it thinks it wants to launch, like we do in the Python wrapper today. Once we do this, we can reasonably get rid of one of the 3 modes.

## Alternatives

I spent a lot of time trying to get test classloading to work with a single approach for all versions, but alas, I don't believe there is one. There are a couple of subtle requirements, which cannot be easily satisfied, including:
- `org.junit.Test` needs to be in the same classloader chain as the test class itself, due to `BlockJUnit4ClassRunner` directly relying on `Test.class`
- `jimfs` needs to be in system classloader, and so do its dependencies, due to the way protocol loading works in the JCL

Some other approaches considered:
- For both Java 8- and Java 9+, make the test classloader a child of the system classloader, with the `CompilerTreeRunner` classloader a child of the compiler classloader
  - Problem: the test class ends up in the `CompilerTreeRunner` classloader, but the JUnit stuff ends up in the test classloader, and these need to be in the same parent chain
- Make the test classloader a child of the compiler classloader
  - Problem: plugin stuff still needs to be in the system classloader
- Java 8- remains the same, but for Java 9+, the test classloader is a child of the system classloader
  - Problem: jimfs, and its dependencies, including Guava, need to be in the system classloader

Reviewed By: jkeljo

shipit-source-id: 365b21469e